had to look up the solution to finish
need to take follow up notes on filter(), forEach, ...,
includes(), and the ! operator when used with methods. Super hard
would not have been able to complete without looking at the
solution. Was thinking of turning the array into a string, then
removing the given items to remove, then putting the string back
into an array.